Address replay extraction review follow-ups
